requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
aiohttp
faster-whisper
torch
sentencepiece
python-dotenv

